backend/requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
# Use PyMySQL on Windows to avoid compiling mysqlclient C extensions
PyMySQL==1.0.3
python-dotenv==1.0.0
google-generativeai==0.3.0
protobuf==3.20.3
python-magic==0.4.27

# Cryptography is required for MySQL caching_sha2_password auth via PyMySQL
cryptography>=42.0.5

# Pin setuptools to avoid pkg_resources deprecation warnings in some packages
# pkg_resources is slated for removal in newer setuptools; pin to <81 until
# dependent packages are updated. See runserver.err warnings.
setuptools<81